+++ src/index.ts
@@ -6,7 +6,10 @@ export * from "./PrintOne";
 
 // Models
 export * from "./models/Address";
+export * from "./models/Batch";
 export * from "./models/Company";
+export * from "./models/Coupon";
+export * from "./models/CouponCode";
 export * from "./models/CsvOrder";
 export * from "./models/CustomFile";
 export * from "./models/Order";
@@ -14,7 +17,6 @@ export * from "./models/PaginatedResponse";
 export * from "./models/Preview";
 export * from "./models/PreviewDetails";
 export * from "./models/Template";
-export * from "./models/Batch";
 
 // Enums
 export * from "./enums/CsvStatus";